almost make UPDATE/INSERT statements work
Still doesn't work for querying from VIEWs, which is
important. Views don't "show" the PKs of the source
tables, which makes it diffult to UPDATE/INSERT based
on PK of course. I've got 2 solutions in mind: let
the user specify the desired PK, or just take the
PK from the PostgreSQL side. The second solution
could also allow an override from the user, of course.